move functions to app
