+++ src/components/App/AddQuestions/AddQuestions.jsx
@@ -2,107 +2,107 @@ import React, { Component } from 'react';
 import './AddQuestions.css';
 
 export default class AddQuestion extends Component {
-    constructor(state) {
-      super();
-      this.state = {
-        newQuestion: '',
-        newCorrectAnswer: '',
-        newAnswerA: '',
-        newAnswerB: '',
-        newAnswerC: '',
-        newAnswerD: '',
-        newCategory: '',
-        newDifficulty: '',
-        createdBy: '',
-      };
-    }
+    // constructor(state) {
+    //   super();
+    //   this.state = {
+    //     newQuestion: '',
+    //     newCorrectAnswer: '',
+    //     newAnswerA: '',
+    //     newAnswerB: '',
+    //     newAnswerC: '',
+    //     newAnswerD: '',
+    //     newCategory: '',
+    //     newDifficulty: '',
+    //     createdBy: '',
+    //   };
+    // }
 
-    updateNewQuestion(e) {
-      this.setState({
-        newQuestion: e.target.value,
-      })
-    }
+    // updateNewQuestion(e) {
+    //   this.setState({
+    //     newQuestion: e.target.value,
+    //   })
+    // }
 
-    updateNewCorrectAnswer(e) {
-      this.setState({
-        newCorrectAnswer: e.target.value,
-      })
-    }
+    // updateNewCorrectAnswer(e) {
+    //   this.setState({
+    //     newCorrectAnswer: e.target.value,
+    //   })
+    // }
 
-    updateNewAnswerA(e) {
-      this.setState({
-        newAnswerA: e.target.value,
-      })
-    }
+    // updateNewAnswerA(e) {
+    //   this.setState({
+    //     newAnswerA: e.target.value,
+    //   })
+    // }
 
-    updateNewAnswerB(e) {
-      this.setState({
-        newAnswerB: e.target.value,
-      })
-    }
+    // updateNewAnswerB(e) {
+    //   this.setState({
+    //     newAnswerB: e.target.value,
+    //   })
+    // }
 
-    updateNewAnswerC(e) {
-      this.setState({
-        newAnswerC: e.target.value,
-      })
-    }
+    // updateNewAnswerC(e) {
+    //   this.setState({
+    //     newAnswerC: e.target.value,
+    //   })
+    // }
 
-    updateNewAnswerD(e) {
-      this.setState({
-        newAnswerD: e.target.value,
-      })
-    }
+    // updateNewAnswerD(e) {
+    //   this.setState({
+    //     newAnswerD: e.target.value,
+    //   })
+    // }
 
-    updateNewCategory(e) {
-      this.setState({
-        newCategory: e.target.value,
-      })
-    }
+    // updateNewCategory(e) {
+    //   this.setState({
+    //     newCategory: e.target.value,
+    //   })
+    // }
 
-    updateNewDifficulty(e) {
-      this.setState({
-        newDifficulty: e.target.value,
-      })
-    }
+    // updateNewDifficulty(e) {
+    //   this.setState({
+    //     newDifficulty: e.target.value,
+    //   })
+    // }
 
-    updateCreatedBy(e) {
-      this.setState({
-        createdBy: e.target.value,
-      })
-    }
+    // updateCreatedBy(e) {
+    //   this.setState({
+    //     createdBy: e.target.value,
+    //   })
+    // }
 
-    submitQuestion() {
-      console.log('adding question!');
-      fetch('/questions', {
-        headers: {
-          'Content-type': 'application/json; charset=UTF-8',
-        },
-        method: 'POST',
-        body: JSON.stringify({
-          question: this.state.newQuestion,
-          correct_answer: this.state.newCorrectAnswer,
-          answerA: this.state.newAnswerA,
-          answerB: this.state.newAnswerB,
-          answerC: this.state.newAnswerC,
-          answerD: this.state.newAnswerD,
-          category: this.state.newCategory,
-          difficulty: this.state.newDifficulty,
-          created_by: this.state.createdBy,
-        })
-      })
-      .then(this.setState({
-          newQuestion: '',
-          newCorrectAnswer: '',
-          newAnswerA: '',
-          newAnswerB: '',
-          newAnswerC: '',
-          newAnswerD: '',
-          newCategory: '',
-          newDifficulty: '',
-          createdBy: '',
-      }))
-      .catch(err => console.log(err));
-    }
+    // submitQuestion() {
+    //   console.log('adding question!');
+    //   fetch('/questions', {
+    //     headers: {
+    //       'Content-type': 'application/json; charset=UTF-8',
+    //     },
+    //     method: 'POST',
+    //     body: JSON.stringify({
+    //       question: this.state.newQuestion,
+    //       correct_answer: this.state.newCorrectAnswer,
+    //       answerA: this.state.newAnswerA,
+    //       answerB: this.state.newAnswerB,
+    //       answerC: this.state.newAnswerC,
+    //       answerD: this.state.newAnswerD,
+    //       category: this.state.newCategory,
+    //       difficulty: this.state.newDifficulty,
+    //       created_by: this.state.createdBy,
+    //     })
+    //   })
+    //   .then(this.setState({
+    //       newQuestion: '',
+    //       newCorrectAnswer: '',
+    //       newAnswerA: '',
+    //       newAnswerB: '',
+    //       newAnswerC: '',
+    //       newAnswerD: '',
+    //       newCategory: '',
+    //       newDifficulty: '',
+    //       createdBy: '',
+    //   }))
+    //   .catch(err => console.log(err));
+    // }
 
   render() {
     return(
@@ -112,67 +112,67 @@ export default class AddQuestion extends Component {
           <input
             type="text"
             placeholder="Create Question"
-            value={this.props.newQuestion}
-            onChange={event => this.updateNewQuestion(event)}
+            value={this.props.state.newQuestion}
+            onChange={this.props.updateNewQuestion}
           />
 
           <input
             type="text"
             placeholder="Enter correct answer"
-            value={this.props.newCorrectAnswer}
-            onChange={event => this.updateNewCorrectAnswer(event)}
+            value={this.props.state.newCorrectAnswer}
+            onChange={this.props.updateNewCorrectAnswer}
           />
 
           <input
             type="text"
             placeholder="Enter first answer choice"
-            value={this.props.newAnswerA}
-            onChange={event => this.updateNewAnswerA(event)}
+            value={this.props.state.newAnswerA}
+            onChange={this.props.updateNewAnswerA}
           />
 
           <input
             type="text"
             placeholder="Enter second answer choice"
-            value={this.props.answerB}
-            onChange={event => this.updateNewAnswerB(event)}
+            value={this.props.state.answerB}
+            onChange={this.props.updateNewAnswerB}
           />
 
           <input
             type="text"
             placeholder="Enter third answer choice"
-            value={this.props.newAnswerC}
-            onChange={event => this.updateNewAnswerC(event)}
+            value={this.props.state.newAnswerC}
+            onChange={this.props.updateNewAnswerC}
           />
 
           <input
             type="text"
             placeholder="Enter fourth answer choice"
-            value={this.props.newAnswerD}
-            onChange={event => this.updateNewAnswerD(event)}
+            value={this.props.state.newAnswerD}
+            onChange={this.props.updateNewAnswerD}
           />
 
           <input
             type="text"
             placeholder="Enter category"
-            value={this.props.newCategory}
-            onChange={event => this.updateNewCategory(event)}
+            value={this.props.state.newCategory}
+            onChange={this.props.updateNewCategory}
           />
 
           <input
             type="text"
             placeholder="Enter difficulty"
-            value={this.props.newDifficulty}
-            onChange={event => this.updateNewDifficulty(event)}
+            value={this.props.state.newDifficulty}
+            onChange={this.props.updateNewDifficulty}
           />
 
           <input
             type="text"
             placeholder="Enter username"
-            value={this.props.createdBy}
-            onChange={event => this.updateCreatedBy(event)}
+            value={this.props.state.createdBy}
+            onChange={this.props.updateCreatedBy}
           />
 
-          <button onClick={event => this.submitQuestion(event)}> Submit Question </button>
+          <button onClick={this.props.submitQuestion}> Submit Question </button>
         </div>
       </container>
       );
